Do not create a new machine-id file
This commit ensures KIWI is not creating a new machine-id empty file
in case it was not provided during the system installation.

Fixes bsc#1141168

(cherry picked from commit afd9b3019ae37980524554452e113d544e9676aa)

+++ kiwi/system/setup.py
@@ -236,8 +236,10 @@ class SystemSetup(object):
         machine_id = os.sep.join(
             [self.root_dir, 'etc', 'machine-id']
         )
-        with open(machine_id, 'w'):
-            pass
+
+        if os.path.exists(machine_id):
+            with open(machine_id, 'w'):
+                pass
 
     def setup_permissions(self):
         """
